+++ src/main/java/subway/common/exception/AuthorizationException.java
@@ -1,13 +1,10 @@
-package subway.exceptions;
+package subway.common.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.UNAUTHORIZED)
 public class AuthorizationException extends RuntimeException {
-    public AuthorizationException() {
-    }
-
     public AuthorizationException(String message) {
         super(message);
     }
