+++ src/main/java/subway/common/exception/NotExistException.java
@@ -1,13 +1,10 @@
-package subway.exceptions;
+package subway.common.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 public class NotExistException extends RuntimeException{
-    public NotExistException() {
-    }
-
     public NotExistException(String message) {
         super(message);
     }
